replace int with int64 for ID and AuthDate attributes in Credentials

+++ tgverifier.go
@@ -13,12 +13,12 @@ var ErrInvalidCreds = errors.New("invalid telegram creds")
 
 // Credentials are Telegram Login credentials available for parsing from JSON.
 type Credentials struct {
-	ID        int    `json:"id"`
+	ID        int64  `json:"id"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
 	Username  string `json:"username"`
 	PhotoURL  string `json:"photo_url"`
-	AuthDate  int    `json:"auth_date"`
+	AuthDate  int64  `json:"auth_date"`
 	Hash      string `json:"hash"`
 }
